123
123

+++ Socket/Assets/Script/GameEngine/Point.cs
@@ -3,7 +3,7 @@ namespace GameEngine
 {
     public class Point
     {
-        public int x;
-        public int y;
+        public float x;
+        public float y;
     }
 }
